refactor: project sharing feature and added permission policies

+++ backend/app/Http/Controllers/ProjectController.php
@@ -4,37 +4,37 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-use App\Http\Resources\ProjectResource;
 use App\Models\Project;
-use App\Models\SharedProject;
+use App\Enums\ProjectUserRole;
+use App\Http\Resources\ProjectResource;
 
 class ProjectController extends Controller
 {
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Project::class);
+
         /// get user from request
         $user = $request->user();
 
-        // fetch projects
-        $projects = Project::where('user_id', $user->id);
+        // fetch projects related to user [owner, editor or viewer]
+        $projects = $user->projects()->with(['users']);
 
-        // fetch shared projects
-        $sharedProjects = SharedProject::where('user_id', $user->id)->pluck('project_id');
-        $sharedProjects = Project::whereIn('id', $sharedProjects);
-
-        $projects = $projects->union($sharedProjects)->orderBy('created_at', 'desc')->simplePaginate(6);
-        $data = ProjectResource::collection($projects);
+        // apply ordering and pagination
+        $projects = $projects->orderBy('created_at', 'desc')->simplePaginate(6);
 
         return response()->json([
-            'data' => $data,
+            'data' => ProjectResource::collection($projects),
             'meta' => [
                 "has_more_pages" => $projects->hasMorePages(),
-            ]
+            ],
         ], 200);
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Project::class);
+
         // validate the request
         $request->validate([
             'title' => 'required|string|max:255',
@@ -45,25 +45,33 @@ class ProjectController extends Controller
 
         // create project
         $project = Project::create([
-            'user_id' => $user->id,
             'title' => $request->title,
         ]);
 
-        $data = new ProjectResource($project);
+        // assign project to user and make owner
+        $project->users()->attach($user->id, ["role" => ProjectUserRole::OWNER]);
 
-        return response()->json(['data' => $data, 'message' => 'Project created'], 201);
+        return response()->json([
+            'data' => new ProjectResource($project),
+            'message' => 'Project created'
+        ], 201);
     }
 
 
     public function show(Project $project)
     {
-        $data = new ProjectResource($project);
-        return response()->json(['data' => $data]);
+        $this->authorize('view', $project);
+
+        return response()->json([
+            'data' => new ProjectResource($project)
+        ]);
     }
 
 
     public function update(Request $request, Project $project)
     {
+        $this->authorize('update', $project);
+
         // validate the request
         $request->validate([
             'title' => 'sometimes|string|max:255',
@@ -71,16 +79,22 @@ class ProjectController extends Controller
 
         // update project
         $project->update($request->all());
-        $data = new ProjectResource($project);
 
-        return response()->json(['data' => $data, 'message' => 'Project updated']);
+        return response()->json([
+            'data' => new ProjectResource($project),
+            'message' => 'Project updated'
+        ]);
     }
 
-    public function destroy(Project $project)
+    public function destroy(Request $request, Project $project)
     {
+        $this->authorize('delete', $project);
+
         // delete project
         $project->delete();
 
-        return response()->json(['message' => 'Project deleted'], 203);
+        return response()->json([
+            'message' => 'Project deleted'
+        ], 203);
     }
 }
